Build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@65696 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DrvHostALSAAudio.cpp
@@ -1664,7 +1664,7 @@ static DECLCALLBACK(uint32_t) drvHostALSAAudioStreamGetWritable(PPDMIHOSTAUDIO p
     snd_pcm_sframes_t cFramesAvail;
     int rc = alsaStreamGetAvail(pStreamALSA->phPCM, &cFramesAvail);
     if (   RT_SUCCESS(rc)
-        && cFramesAvail >= pStreamALSA->Out.cSamplesMin)
+        && (uint32_t)cFramesAvail >= pStreamALSA->Out.cSamplesMin)
     {
         cbAvail = PDMAUDIOSTREAMCFG_S2B(pStreamALSA->pCfg, cFramesAvail);
     }
